ca recovery option
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@9058 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ connect/org.wcs.smart.connect.server/src/main/java/org/wcs/smart/connect/uploader/sync/ChangeLogManager.java
@@ -347,7 +347,7 @@ public enum ChangeLogManager {
 			ChangeLogItem ci = new ChangeLogItem();
 			
 			ci.setUuid((UUID)data[i++]);
-			ci.setRevision( ((BigInteger)data[i++]).longValue() );
+			ci.setRevision( (Long)data[i++] );
 			ci.setAction(Action.valueOf((String)data[i++]));
 			ci.setFileName((String)data[i++]);
 			ci.setTableName((String)data[i++]);
